requirements.txt
streamlit==1.35.0
pillow==10.3.0
fal-client>=0.4.1
requests==2.32.3
python-dotenv==1.0.1
nest_asyncio>=1.5.0